% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{equilibrium_init_create}
\alias{equilibrium_init_create}
\title{Create equilibrium initial conditions}
\usage{
equilibrium_init_create(par)
}
\arguments{
\item{par}{A list of parameters}
}
\value{
A data frame of equilibrium conditions
}
\description{
Create equilibrium initial conditions
}
