% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{phi_eir_rel}
\alias{phi_eir_rel}
\title{Generate equilibrium parameters based on EIR and ft}
\usage{
phi_eir_rel(
  EIR,
  ft,
  ton = 5000,
  toff = 50000,
  init_res = 0.01,
  res_time = 3000,
  rT_r_true = 0.2,
  day0_res = 0.01,
  resistance_trans_mult = 1,
  resistance_dur_mult = 1
)
}
\arguments{
\item{EIR}{Numeric. The Entomological Inoculation Rate.}

\item{ft}{Numeric. The treatment rate.}

\item{ton}{Time at which treatment is turned on}

\item{toff}{Time at which treatment is turned off}

\item{init_res}{Initial resistance level at res_time}

\item{res_time}{Time at which resistance is introduced}

\item{rT_r_true}{True treatment rate for resistant parasites}

\item{day0_res}{Resistant at Day 0. Default = 0}

\item{resistance_trans_mult}{transmission multiplier for resistant parasites (default = 1)}

\item{resistance_dur_mult}{duration multiplier for resistant infections (default = 1)}
}
\value{
A list of generated parameters.
}
\description{
Generate equilibrium parameters based on EIR and ft
}
