% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model.R
\name{plot_model}
\alias{plot_model}
\title{Plot Malaria Model Results}
\usage{
plot_model(model_results, output_vars = NULL)
}
\arguments{
\item{model_results}{A data frame containing the model results or an odin model object.}

\item{output_vars}{A character vector specifying the output variables to plot. If NULL, all output variables will be plotted.}
}
\value{
A ggplot object.
}
\description{
This function plots the results of a malaria model simulation.
}
