% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equilibrium.R
\name{starting_params}
\alias{starting_params}
\title{Generate starting parameters for a range of EIR and ft values}
\usage{
starting_params(
  EIRs = c(0.25, 0.5, 0.75, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, 5, 6, 7, 8, 9, 10, 15, 20,
    30, 50, 100, 200),
  fts = seq(0.1, 0.9, 0.1)
)
}
\arguments{
\item{EIRs}{Numeric vector. The range of Entomological Inoculation Rates.}

\item{fts}{Numeric vector. The range of treatment rates.}
}
\value{
A data frame of starting parameters for each combination of EIR and ft.
}
\description{
Generate starting parameters for a range of EIR and ft values
}
